Fix timezone related bug in visualization script
When running the visualization script the timestamps would be
interpreted in the locality where the script was being run.
Since the script may be run in a timezone that is not the one
from the Potsdam based mensa this meant that the resulting
normalized times would be incorrect.
Fixed by making the Potsdam timezone explicit.