Broken: Aborted echart attempt
This is a lot of WIP code that I want in the log in case we ever
re-attempt to use echarts. The problem with the go-echarts library
is that we just can't get it to work: Settings that we pass
in don't have their intended effects, and at this point
it's a reasonable call to find out whether we can do what we want
to do more quickly in a different library.