Fix bug where y labels would be cut off
This includes changing our dependencies: The last go-echarts release
happened quite long ago, and for this fix we need echart features
which go-echart doesn't yet support. This also changes a whole
bunch of defaults (e.g. showSymbol:false, which we don't want)
Not all of these defaults are currently supported in go-echarts,
so this'll have to do for now.

The alternative would be doing more custom stuff, but that's
currently considered out of scope