Make db query for historical data DST aware
This was interesting. The timestamps we store in the DB are unix
timestamps, but the time intervals we care about are in the
Europe/Berlin timezone. A report from 11:00 Europe/Berlin
can correspond to 12:00 UTC or 13:00 UTC, depending on the
state of daylight saving time. SQLite doesn't currently have an
awareness for this, so we query more data than necessary, and
filter things further in Go.

It's true what they said. Timezones are hard.