Add draft for "initial message" functionality
This includes renaming DB fields and corresponding code from
"seconds" to "UTCMinutes" to be clearer, as well as the actual
code/queries for the functionality.

"Initial message" is the feature for feature equivalence with
other mensa bots: When your window of interest opens you receive
an initial message. I'm not happy with the naming, but it is the
best I could come up with for now