Add backend code to display correct settings for user
We do this by passing query parameters in the URL. The alternative
would be to write up an endpoint that serves it as json, and the
web html/javascript requests it via AJAX or something.

THe choice to not do that was very much gut feeling. I justify it
as follows
- One roundtrip less before showing actual settings
- Opening up our server to non-telegram requests might come with
bigger rebuilds

Adn that's it, really. In the end, I think it's a choice that doesn't
really matter.